WIP - debugging contour with separate function
Debugging contour now done with separate function.

Also this time I make a random 2D distribution of points classified as 1 when inside a radius of 0.5, 0 otherwise, and try to plot the contour for it. I am expecting a circle of radius 0.5. However, something is wrong with the contour z value. 

Generalized the removal of outliers for gaussian distributions. 
Generalized the plot limits.